# backend/src/main/java/itmo/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Infrastructure Management API
  description: API для управления виртуальными машинами, деплоями, снапшотами, мониторингом и drift-отчётами
  version: 1.0.0
  contact:
    name: Kostik Bushmelev
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://144.31.219.63:8080
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/vms/{vmId}/drift:
    post:
      tags: [Drift]
      summary: Запустить проверку drift для VM
      parameters:
        - $ref: '#/components/parameters/VmIdParam'
      responses:
        '201':
          description: Проверка запущена, отчёт создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftReportResponse'
        '404':
          description: VM не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/vms/{vmId}/reference-drift:
    post:
      tags: [Drift]
      summary: Сравнить VM с эталонным snapshot
      parameters:
        - $ref: '#/components/parameters/VmIdParam'
      responses:
        '201':
          description: Сравнение с эталоном выполнено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftReportResponse'
        '404':
          description: VM или эталонный snapshot не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/drift-reports:
    get:
      tags: [Drift]
      summary: Все drift-отчёты
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DriftStatus'
        - name: vmId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список отчётов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDriftReportResponse'
  /api/v1/drift-reports/{id}:
    get:
      tags: [Drift]
      summary: Детали drift-отчёта
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Отчёт найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftReportResponse'
        '404':
          description: Не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ======================== DEPLOYMENTS ========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VmIdParam:
      name: vmId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    DriftStatus:
      type: string
      enum: [CLEAN, DRIFTED, UNKNOWN]
...
    DriftReportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vmId:
          type: string
          format: uuid
        vmName:
          type: string
        status:
          $ref: '#/components/schemas/DriftStatus'
        expectedState:
          type: object
          description: JSON ожидаемого состояния из OpenTofu state
          additionalProperties: true
        actualState:
          type: object
          description: JSON фактического состояния VM
          additionalProperties: true
        differences:
          type: array
          description: Список отличий
          items:
            $ref: '#/components/schemas/DriftDifference'
        checkedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    DriftDifference:
      type: object
      properties:
        field:
          type: string
          example: memoryMb
        expected:
          type: string
          example: "1024"
        actual:
          type: string
          example: "2048"

    # ======================== SERVICE ========================
...
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    PageDriftReportResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DriftReportResponse'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Virtual machine with id 123 not found
        path:
          type: string
          example: /api/v1/vms/123